refactor(test): extract domain layer tests to tests/unit/
Move all inline #[cfg(test)] modules from domain entities, value objects,
and repository interfaces to external test files under tests/unit/.
All tests use only public API (formix:: imports). Repository mod.rs tests
(9 tests using private TestEntity) were deleted as they required private access.

Moved: 78 tests (51 domain + 29 repository - 9 deleted = 71 moved + 7 new mock tests)
