+++ client/src/repositories/mod.rs
@@ -90,187 +90,3 @@ pub trait Repository<T, ID> {
     /// Returns only the entities that were found. Missing IDs are silently ignored.
     async fn find_by_ids(&self, ids: &[ID]) -> DomainResult<Vec<T>>;
 }
-
-#[cfg(test)]
-#[allow(clippy::indexing_slicing)]
-mod tests {
-    use super::*;
-    use std::collections::HashMap;
-    use std::sync::RwLock;
-
-    /// Simple entity for testing
-    #[derive(Clone, Debug, PartialEq)]
-    struct TestEntity {
-        id: String,
-        data: String,
-    }
-
-    /// Mock repository implementation for testing base Repository trait
-    struct MockRepository {
-        storage: RwLock<HashMap<String, TestEntity>>,
-    }
-
-    impl MockRepository {
-        fn new() -> Self {
-            Self {
-                storage: RwLock::new(HashMap::new()),
-            }
-        }
-    }
-
-    #[async_trait]
-    impl Repository<TestEntity, String> for MockRepository {
-        async fn save(&self, entity: &TestEntity) -> DomainResult<()> {
-            self.storage
-                .write()
-                .unwrap()
-                .insert(entity.id.clone(), entity.clone());
-            Ok(())
-        }
-
-        async fn find_by_id(&self, id: &String) -> DomainResult<Option<TestEntity>> {
-            Ok(self.storage.read().unwrap().get(id).cloned())
-        }
-
-        async fn delete(&self, id: &String) -> DomainResult<()> {
-            self.storage.write().unwrap().remove(id);
-            Ok(())
-        }
-
-        async fn exists(&self, id: &String) -> DomainResult<bool> {
-            Ok(self.storage.read().unwrap().contains_key(id))
-        }
-
-        async fn find_by_ids(&self, ids: &[String]) -> DomainResult<Vec<TestEntity>> {
-            let storage = self.storage.read().unwrap();
-            Ok(ids
-                .iter()
-                .filter_map(|id| storage.get(id).cloned())
-                .collect())
-        }
-    }
-
-    /// Compile-time verification that Repository requires Send + Sync
-    fn assert_send_sync<T: Send + Sync>() {}
-
-    #[test]
-    fn mock_repository_is_send_sync() {
-        assert_send_sync::<MockRepository>();
-    }
-
-    #[tokio::test]
-    async fn test_find_by_id_not_found() {
-        let repo = MockRepository::new();
-        let result = repo.find_by_id(&"nonexistent".to_string()).await.unwrap();
-        assert!(result.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_save_and_find_by_id() {
-        let repo = MockRepository::new();
-        let entity = TestEntity {
-            id: "test-1".to_string(),
-            data: "test data".to_string(),
-        };
-
-        repo.save(&entity).await.unwrap();
-        let found = repo.find_by_id(&"test-1".to_string()).await.unwrap();
-
-        assert_eq!(found, Some(entity));
-    }
-
-    #[tokio::test]
-    async fn test_exists() {
-        let repo = MockRepository::new();
-        let entity = TestEntity {
-            id: "test-1".to_string(),
-            data: "test data".to_string(),
-        };
-
-        assert!(!repo.exists(&"test-1".to_string()).await.unwrap());
-        repo.save(&entity).await.unwrap();
-        assert!(repo.exists(&"test-1".to_string()).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn test_delete() {
-        let repo = MockRepository::new();
-        let entity = TestEntity {
-            id: "test-1".to_string(),
-            data: "test data".to_string(),
-        };
-
-        repo.save(&entity).await.unwrap();
-        assert!(repo.exists(&"test-1".to_string()).await.unwrap());
-
-        repo.delete(&"test-1".to_string()).await.unwrap();
-        assert!(!repo.exists(&"test-1".to_string()).await.unwrap());
-    }
-
-    #[tokio::test]
-    async fn test_delete_nonexistent_succeeds() {
-        let repo = MockRepository::new();
-        // Deleting non-existent entity should not error
-        let result = repo.delete(&"nonexistent".to_string()).await;
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn test_find_by_ids_batch_retrieval() {
-        let repo = MockRepository::new();
-        let entities = vec![
-            TestEntity {
-                id: "test-1".to_string(),
-                data: "data 1".to_string(),
-            },
-            TestEntity {
-                id: "test-2".to_string(),
-                data: "data 2".to_string(),
-            },
-            TestEntity {
-                id: "test-3".to_string(),
-                data: "data 3".to_string(),
-            },
-        ];
-
-        for entity in &entities {
-            repo.save(entity).await.unwrap();
-        }
-
-        let ids = vec!["test-1".to_string(), "test-3".to_string()];
-        let found = repo.find_by_ids(&ids).await.unwrap();
-
-        assert_eq!(found.len(), 2);
-        assert!(found.iter().any(|e| e.id == "test-1"));
-        assert!(found.iter().any(|e| e.id == "test-3"));
-    }
-
-    #[tokio::test]
-    async fn test_find_by_ids_with_missing_ids() {
-        let repo = MockRepository::new();
-        let entity = TestEntity {
-            id: "test-1".to_string(),
-            data: "data 1".to_string(),
-        };
-        repo.save(&entity).await.unwrap();
-
-        // Request includes both existing and non-existing IDs
-        let ids = vec![
-            "test-1".to_string(),
-            "nonexistent".to_string(),
-            "also-missing".to_string(),
-        ];
-        let found = repo.find_by_ids(&ids).await.unwrap();
-
-        // Only the existing entity should be returned
-        assert_eq!(found.len(), 1);
-        assert_eq!(found[0].id, "test-1");
-    }
-
-    #[tokio::test]
-    async fn test_find_by_ids_empty_input() {
-        let repo = MockRepository::new();
-        let found = repo.find_by_ids(&[]).await.unwrap();
-        assert!(found.is_empty());
-    }
-}
