+++ client/tests/unit/domain/entities/mod.rs
@@ -1,0 +1,5 @@
+mod capsule;
+mod cfrag;
+mod kfrag;
+mod secret;
+mod share;
