+++ client/tests/unit/domain/mod.rs
@@ -1,0 +1,2 @@
+mod entities;
+mod value_objects;
