+++ client/tests/unit/domain/value_objects/mod.rs
@@ -1,0 +1,4 @@
+mod ids;
+mod key_pair;
+mod secret_data;
+mod symmetric_key;
