+++ client/tests/unit/helpers/mod.rs
@@ -1,0 +1,1 @@
+// Common test helpers and mock utilities for unit tests
