+++ client/tests/unit/repositories/mod.rs
@@ -1,0 +1,5 @@
+mod capsule_interface;
+mod cfrag_interface;
+mod kfrag_interface;
+mod secret_interface;
+mod share_interface;
